@@ -15,22 +15,23 @@
   "dependencies": {
     "@angular/animations": "^4.3.4",
     "@angular/cdk": "^2.0.0-beta.8",
-    "@angular/common": "^4.0.0",
-    "@angular/compiler": "^4.0.0",
-    "@angular/core": "^4.0.0",
+    "@angular/common": "^4.3.4",
+    "@angular/compiler": "^4.3.4",
+    "@angular/core": "^4.3.4",
     "@angular/flex-layout": "^2.0.0-beta.8",
-    "@angular/forms": "^4.0.0",
-    "@angular/http": "^4.0.0",
+    "@angular/forms": "^4.3.4",
+    "@angular/http": "^4.3.4",
     "@angular/material": "^2.0.0-beta.8",
-    "@angular/platform-browser": "^4.0.0",
-    "@angular/platform-browser-dynamic": "^4.0.0",
-    "@angular/router": "^4.0.0",
+    "@angular/platform-browser": "^4.3.4",
+    "@angular/platform-browser-dynamic": "^4.3.4",
+    "@angular/router": "^4.3.4",
     "axios": "^0.16.2",
     "body-parser": "^1.17.2",
     "cookie-parser": "^1.4.3",
     "core-js": "^2.4.1",
     "dotenv": "^4.0.0",
     "express-session": "^1.15.4",
+    "font-awesome": "^4.7.0",
     "hammerjs": "^2.0.8",
     "mongoose": "^4.11.3",
     "passport": "^0.3.2",
@@ -41,7 +42,7 @@
   },
   "devDependencies": {
     "@angular/cli": "1.2.1",
-    "@angular/compiler-cli": "^4.0.0",
+    "@angular/compiler-cli": "^4.3.4",
     "@angular/language-service": "^4.0.0",
     "@types/jasmine": "~2.5.53",
     "@types/jasminewd2": "~2.0.2",
